mat dialog

+++ src/app/admin/tags/tags.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
-import {FormControl, Validators} from '@angular/forms';
-import {Observable} from 'rxjs';
-import {TagsService} from 'src/app/services/tags.service';
+import { MatDialog } from '@angular/material/dialog';
+import { Observable } from 'rxjs';
+import { ConfirmationDialogComponent } from 'src/app/dialogs/confirmation-dialog/confirmation-dialog.component';
+import { TagsDialogComponent } from 'src/app/dialogs/tags-dialog/tags-dialog.component';
+import { TagsService } from 'src/app/services/tags.service';
 
 @Component({
   selector: 'app-tags',
@@ -9,46 +11,53 @@ import {TagsService} from 'src/app/services/tags.service';
   styleUrls: ['./tags.component.css']
 })
 export class TagsComponent implements OnInit {
-
-	tags$!: Observable<any[]>;
-
-	tagControl: FormControl = new FormControl('', [Validators.required])
+  tags$!: Observable<any[]>;
 
   constructor(
-		private _tagsService: TagsService
-	) { }
+    private _tagsService: TagsService,
+    private _matDialog: MatDialog
+  ) {}
 
   ngOnInit(): void {
-		this._getTags();
+    this._getTags();
   }
 
-	private _getTags() {
-		this.tags$ = this._tagsService.getAllTag();
-	}
+  private _getTags() {
+    this.tags$ = this._tagsService.getAllTag();
+  }
 
-	addTag() {
-		const data = {
-			title: this.tagControl.value 
-		}
-		this._tagsService.addTag(data).subscribe(
+  addTag(tag?: any) {
+    const dialog = this._matDialog.open(TagsDialogComponent, {
+      width: '440px',
+			data: tag
+    });
+
+		dialog.afterClosed().subscribe(
 			res => {
-				this._getTags();
-				console.log(res);
-			},
-			err => {
-				console.log(err);
+				if (res) {
+					this._getTags();
+				}
 			}
 		)
-	}
+  }
 
-	delete(tagId: string) {
-		this._tagsService.deleteTag(tagId).subscribe(
-			() => {
-				this._getTags();
-			},
-			err => {
-				console.log(err)
-			}
-		)
-	}
+  delete(tagId: string) {
+    const dialog = this._matDialog.open(ConfirmationDialogComponent, {
+      width: '440px'
+    });
+
+    dialog.afterClosed().subscribe(res => {
+      if (res) {
+        this._tagsService.deleteTag(tagId).subscribe(
+          () => {
+            this._getTags();
+          },
+          err => {
+            console.log(err);
+          }
+        );
+      }
+      console.log(res);
+    });
+  }
 }
